+++ src/plugin/manager/loadbalancer/healthmonitors_manager.py
@@ -2,7 +2,7 @@ import logging
 from spaceone.inventory.plugin.collector.lib import *
 
 from plugin.conf.cloud_service_conf import AUTH_TYPE, REGION, ASSET_URL
-from plugin.connector.loadbalancer.healthmonitors_connector import LBHealthMonitorsConnector
+from plugin.connector.loadbalancer.healthmonitor_connector import LBHealthMonitorsConnector
 from plugin.manager.base import NHNCloudBaseManager
 
 _LOGGER = logging.getLogger("cloudforet")
