+++ src/plugin/manager/network_acl/binding_manager.py
@@ -1,24 +1,24 @@
 import logging
 from spaceone.inventory.plugin.collector.lib import *
 
-from plugin.connector.block_storage.block_storage_type_connector import BlockStorageTypeConnector
+from plugin.connector.network_acl.binding_connector import NACLBindingsConnector
 from plugin.conf.cloud_service_conf import AUTH_TYPE, REGION, ASSET_URL
 from plugin.manager.base import NHNCloudBaseManager
 
 _LOGGER = logging.getLogger("cloudforet")
 
 
-class BlockStorageTypeManager(NHNCloudBaseManager):
+class BindingManager(NHNCloudBaseManager):
     auth_type = AUTH_TYPE.TOKEN
-    AVAILABLE_REGIONS = [REGION.KR1, REGION.KR2, REGION.JP1]
+    AVAILABLE_REGIONS = [REGION.KR1, REGION.KR2]
 
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
 
-        self.cloud_service_group = "Block Storage"
-        self.cloud_service_type = "Block Storage Type"
+        self.cloud_service_group = "Network ACL"
+        self.cloud_service_type = "Network ACL Binding"
         self.provider = "nhncloud"
-        self.metadata_path = "metadata/block_storage/block_storage_type.yaml"
+        self.metadata_path = "metadata/network_acl/binding.yaml"
 
     def create_cloud_service_type(self):
         cloud_service_type = make_cloud_service_type(
@@ -29,23 +29,23 @@ class BlockStorageTypeManager(NHNCloudBaseManager):
             is_primary=True,
             is_major=True,
             tags={
-                "spaceone:icon": f"{ASSET_URL}/block_storage.png"
+                "spaceone:icon": f"{ASSET_URL}/network_acl.png"
             },
-            labels=["Management"]
+            labels=["Networking", "Security"]
         )
 
         return cloud_service_type
 
     def create_cloud_service(self, secret_data):
-        block_connector = BlockStorageTypeConnector()
+        bindings_connector = NACLBindingsConnector()
         for AVAILABLE_REGION in self.AVAILABLE_REGIONS:
-            resources = block_connector.get_types(secret_data, AVAILABLE_REGION)
+            resources = bindings_connector.get_bindings(secret_data, AVAILABLE_REGION)
             for resource in resources:
                 reference = {
                     "resource_id": resource.get("id")
                 }
                 cloud_service = make_cloud_service(
-                    name=resource['name'],
+                    name=resource['id'],
                     cloud_service_type=self.cloud_service_type,
                     cloud_service_group=self.cloud_service_group,
                     provider=self.provider,
@@ -55,3 +55,4 @@ class BlockStorageTypeManager(NHNCloudBaseManager):
                     region_code=AVAILABLE_REGION.name
                 )
                 yield cloud_service
+
