+++ src/plugin/manager/network_acl/network_acl_manager.py
@@ -2,23 +2,23 @@ import logging
 from spaceone.inventory.plugin.collector.lib import *
 
 from plugin.conf.cloud_service_conf import AUTH_TYPE, REGION, ASSET_URL
-from plugin.connector.sg.security_groups_connector import SGConnector
+from plugin.connector.network_acl.network_acl_connector import NACLConnector
 from plugin.manager.base import NHNCloudBaseManager
 
 _LOGGER = logging.getLogger("cloudforet")
 
 
-class SGManager(NHNCloudBaseManager):
+class NetworkACLManager(NHNCloudBaseManager):
     auth_type = AUTH_TYPE.TOKEN
-    AVAILABLE_REGIONS = [REGION.KR1, REGION.KR2, REGION.JP1]
+    AVAILABLE_REGIONS = [REGION.KR1, REGION.KR2]
 
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
 
-        self.cloud_service_group = "Security Group"
-        self.cloud_service_type = "Security Group"
+        self.cloud_service_group = "Network ACL"
+        self.cloud_service_type = "Network ACL"
         self.provider = "nhncloud"
-        self.metadata_path = "metadata/sg/security_group.yaml"
+        self.metadata_path = "metadata/network_acl/network_acl.yaml"
 
     def create_cloud_service_type(self):
         cloud_service_type = make_cloud_service_type(
@@ -29,16 +29,17 @@ class SGManager(NHNCloudBaseManager):
             is_primary=True,
             is_major=True,
             tags={
-                "spaceone:icon": f"{ASSET_URL}/security_group.png"
+                "spaceone:icon": f"{ASSET_URL}/network_acl.png"
             },
-            labels=["Security", "Networking"]
+            labels=["Networking", "Security"]
         )
+
         return cloud_service_type
 
     def create_cloud_service(self, secret_data):
-        sg_connector = SGConnector()
+        nacl_connector = NACLConnector()
         for AVAILABLE_REGION in self.AVAILABLE_REGIONS:
-            resources = sg_connector.get_security_groups(secret_data, AVAILABLE_REGION)
+            resources = nacl_connector.get_nacl(secret_data, AVAILABLE_REGION)
             for resource in resources:
                 reference = {
                     "resource_id": resource.get("id")
@@ -54,3 +55,4 @@ class SGManager(NHNCloudBaseManager):
                     region_code=AVAILABLE_REGION.name
                 )
                 yield cloud_service
+
