+++ src/plugin/manager/network_acl/rule_manger.py
@@ -1,24 +1,24 @@
 import logging
 from spaceone.inventory.plugin.collector.lib import *
 
+from plugin.connector.network_acl.rule_connector import NACLRulesConnector
 from plugin.conf.cloud_service_conf import AUTH_TYPE, REGION, ASSET_URL
-from plugin.connector.instance.instance_AZ_connector import InstanceAZConnector
 from plugin.manager.base import NHNCloudBaseManager
 
 _LOGGER = logging.getLogger("cloudforet")
 
 
-class InstanceAZManager(NHNCloudBaseManager):
+class RuleManager(NHNCloudBaseManager):
     auth_type = AUTH_TYPE.TOKEN
-    AVAILABLE_REGIONS = [REGION.KR1, REGION.KR2, REGION.JP1]
+    AVAILABLE_REGIONS = [REGION.KR1, REGION.KR2]
 
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
 
-        self.cloud_service_group = "Instance"
-        self.cloud_service_type = "Instance Availability Zone"
+        self.cloud_service_group = "Network ACL"
+        self.cloud_service_type = "Network ACL Rule"
         self.provider = "nhncloud"
-        self.metadata_path = "metadata/instance/instance_AZ_resource.yaml"
+        self.metadata_path = "metadata/network_acl/rule.yaml"
 
     def create_cloud_service_type(self):
         cloud_service_type = make_cloud_service_type(
@@ -29,24 +29,23 @@ class InstanceAZManager(NHNCloudBaseManager):
             is_primary=True,
             is_major=True,
             tags={
-                "spaceone:icon": f"{ASSET_URL}/instance.png"
+                "spaceone:icon": f"{ASSET_URL}/network_acl.png"
             },
-            labels=["Compute", "Networking"]
+            labels=["Networking", "Security"]
         )
 
         return cloud_service_type
 
     def create_cloud_service(self, secret_data):
-        resource_connector = InstanceAZConnector()
-
+        rules_connector = NACLRulesConnector()
         for AVAILABLE_REGION in self.AVAILABLE_REGIONS:
-            resources = resource_connector.get_availability_zone(secret_data, AVAILABLE_REGION)
+            resources = rules_connector.get_nacl_rules(secret_data, AVAILABLE_REGION)
             for resource in resources:
                 reference = {
-                    "resource_id": resource.get("zoneName")
+                    "resource_id": resource.get("id")
                 }
                 cloud_service = make_cloud_service(
-                    name=resource["zoneName"],
+                    name=resource['id'],
                     cloud_service_type=self.cloud_service_type,
                     cloud_service_group=self.cloud_service_group,
                     provider=self.provider,
@@ -56,3 +55,4 @@ class InstanceAZManager(NHNCloudBaseManager):
                     region_code=AVAILABLE_REGION.name
                 )
                 yield cloud_service
+
