+++ src/plugin/manager/security_group/rule_manager.py
@@ -1,24 +1,24 @@
 import logging
 from spaceone.inventory.plugin.collector.lib import *
 
-from plugin.connector.nacl.rules_connector import NACLRulesConnector
 from plugin.conf.cloud_service_conf import AUTH_TYPE, REGION, ASSET_URL
+from plugin.connector.security_group.rule_connector import SGRulesConnector
 from plugin.manager.base import NHNCloudBaseManager
 
 _LOGGER = logging.getLogger("cloudforet")
 
 
-class NACLRulesManager(NHNCloudBaseManager):
+class RuleManager(NHNCloudBaseManager):
     auth_type = AUTH_TYPE.TOKEN
-    AVAILABLE_REGIONS = [REGION.KR1, REGION.KR2]
+    AVAILABLE_REGIONS = [REGION.KR1, REGION.KR2, REGION.JP1]
 
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
 
-        self.cloud_service_group = "Network ACL"
-        self.cloud_service_type = "Network ACL Rule"
+        self.cloud_service_group = "Security Group"
+        self.cloud_service_type = "Security Rule"
         self.provider = "nhncloud"
-        self.metadata_path = "metadata/nacl/rules.yaml"
+        self.metadata_path = "metadata/security_group/rule.yaml"
 
     def create_cloud_service_type(self):
         cloud_service_type = make_cloud_service_type(
@@ -29,17 +29,17 @@ class NACLRulesManager(NHNCloudBaseManager):
             is_primary=True,
             is_major=True,
             tags={
-                "spaceone:icon": f"{ASSET_URL}/network_acl.png"
+                "spaceone:icon": f"{ASSET_URL}/security_group.png"
             },
-            labels=["Networking", "Security"]
+            labels=["Security", "Networking"]
         )
 
         return cloud_service_type
 
     def create_cloud_service(self, secret_data):
-        rules_connector = NACLRulesConnector()
+        sg_rules_connector = SGRulesConnector()
         for AVAILABLE_REGION in self.AVAILABLE_REGIONS:
-            resources = rules_connector.get_nacl_rules(secret_data, AVAILABLE_REGION)
+            resources = sg_rules_connector.get_security_rules(secret_data, AVAILABLE_REGION)
             for resource in resources:
                 reference = {
                     "resource_id": resource.get("id")
@@ -55,4 +55,3 @@ class NACLRulesManager(NHNCloudBaseManager):
                     region_code=AVAILABLE_REGION.name
                 )
                 yield cloud_service
-
