+++ src/plugin/manager/security_group/security_group_manager.py
@@ -2,23 +2,23 @@ import logging
 from spaceone.inventory.plugin.collector.lib import *
 
 from plugin.conf.cloud_service_conf import AUTH_TYPE, REGION, ASSET_URL
-from plugin.connector.nacl.nacl_connector import NACLConnector
+from plugin.connector.security_group.security_group_connector import SGConnector
 from plugin.manager.base import NHNCloudBaseManager
 
 _LOGGER = logging.getLogger("cloudforet")
 
 
-class NACLManager(NHNCloudBaseManager):
+class SecurityGroupManager(NHNCloudBaseManager):
     auth_type = AUTH_TYPE.TOKEN
-    AVAILABLE_REGIONS = [REGION.KR1, REGION.KR2]
+    AVAILABLE_REGIONS = [REGION.KR1, REGION.KR2, REGION.JP1]
 
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
 
-        self.cloud_service_group = "Network ACL"
-        self.cloud_service_type = "Network ACL"
+        self.cloud_service_group = "Security Group"
+        self.cloud_service_type = "Security Group"
         self.provider = "nhncloud"
-        self.metadata_path = "metadata/nacl/nacl.yaml"
+        self.metadata_path = "metadata/security_group/security_group.yaml"
 
     def create_cloud_service_type(self):
         cloud_service_type = make_cloud_service_type(
@@ -29,17 +29,16 @@ class NACLManager(NHNCloudBaseManager):
             is_primary=True,
             is_major=True,
             tags={
-                "spaceone:icon": f"{ASSET_URL}/network_acl.png"
+                "spaceone:icon": f"{ASSET_URL}/security_group.png"
             },
-            labels=["Networking", "Security"]
+            labels=["Security", "Networking"]
         )
-
         return cloud_service_type
 
     def create_cloud_service(self, secret_data):
-        nacl_connector = NACLConnector()
+        sg_connector = SGConnector()
         for AVAILABLE_REGION in self.AVAILABLE_REGIONS:
-            resources = nacl_connector.get_nacl(secret_data, AVAILABLE_REGION)
+            resources = sg_connector.get_security_groups(secret_data, AVAILABLE_REGION)
             for resource in resources:
                 reference = {
                     "resource_id": resource.get("id")
@@ -55,4 +54,3 @@ class NACLManager(NHNCloudBaseManager):
                     region_code=AVAILABLE_REGION.name
                 )
                 yield cloud_service
-
