+++ src/plugin/metadata/security_group/security_group.yaml
